% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{pubdata1}
\alias{pd1.count.matrix}
\alias{pd1.similarity}
\alias{pd1.uncat.refs}
\alias{pubdata1}
\title{pubdata1}
\format{\describe{
 \item{\code{pd1.count.matrix}}{the count of referenced disciplines for each publication}
 \item{\code{pd1.uncat.refs}}{the count of referenced disciplines for each publication}
 \item{\code{pd1.similarity}}{between disciplines as given in Porter and Rafols, 2009.}
}}
\usage{
data("pubdata1")
}
\description{
Small example dataset with 5 publications that have most of their references categorized into disciplines.
The dataset contains the following information: A matrix of counts of referenced disciplines for each publication, a vector of counts of uncategorized references in each publication, and a matrix that contains a measure of similarity between disciplines.
}
\references{
Porter, A. and Rafols, I. (2009) Is science becoming more interdisciplinary? Measuring and mapping six research fields over time. Scientometrics, Vol. 81, No. 3 (719-745). DOI:10.1007/s11192-008-2197-2
}
\keyword{datasets}

